+++ addFC_Data.py
@@ -13,6 +13,8 @@ configuration = {
     'unfold_dxf': True,
     'unfold_svg': False,
     'unfold_stp': False,
+    'unfold_centering': False,
+    'unfold_along_y': False,
     'unfold_file_name': 'Index',
     'unfold_file_signature': 'None',
     'unfold_prefix': 'Unfold',