+++ addFC_Unfold.py
@@ -8,18 +8,12 @@ import FreeCAD
 import importDXF
 import ImportGui
 import importSVG
-import math
 import os
 import Part
 import re
 import shutil
 import time
 
-
-if P.afc_additions['sm'][0]:
-    from SheetMetalUnfoldCmd import SMUnfoldUnattendedCommandClass as u
-else:
-    u = None
 
 if P.afc_additions['ezdxf'][0]:
     import ezdxf.filemanagement
@@ -28,36 +22,48 @@ else:
 
 
 REPRODUCTION = 'Reproduction'
-UNFOLD_SKETCH = 'Unfold_Sketch'
-VERIFICATION_SKETCH = 'Unfold_Sketch_Outline'
-UNFOLD_OBJECT = 'Unfold'
 
-GARBAGE = (
-    # case sensitive!
-    REPRODUCTION,
-    UNFOLD_SKETCH,
-    VERIFICATION_SKETCH,
-    'Unfold_Sketch_bends',
-    'Unfold_Sketch_Internal',
-    UNFOLD_OBJECT,  # the object must be the last
-)
+UNFOLD_OBJECT = 'Unfold'
+UNFOLD_SKETCH = 'Unfold_Sketch'
 
 FORBIDDEN = re.escape('<>:"?*/|\\')
 
 
-def add_signature(file, sign: str, width: float, size: int) -> None:
+def dxf_postprocessor(file, sign: str, bb) -> None:
     if ezdxf is None:
         return
     file = ezdxf.filemanagement.readfile(file)
+
+    # cleaning:
+    file.layers.remove('none')
+    file.layers.remove('Defpoints')
+    if sign == '':
+        file.save()
+        return
+
+    # signing:
     model = file.modelspace()
-    x = -width / 2 + size / 2
-    y = -size / 2
-    model.add_text(sign, height=size).set_placement((x, y))
+    x, y = bb.XLength, bb.YLength
+
+    size = int(abs(x) / len(sign))
+    size = max(2, min(size, 200))
+    size_verify = int(abs(y) - 10)
+    if size > size_verify:
+        size = size_verify
+
+    sign_width = len(sign) * (size / 100 * 87)  # theoretical width...
+
+    x = int(bb.Center.x - sign_width / 2)
+    y = int(bb.Center.y - size / 2)
+
+    model.add_text(sign,
+                   height=size,
+                   dxfattribs={'layer': 'Text'}).set_placement((x, y))
     file.save()
 
 
-def cleaning(skip: str) -> None:
-    for i in GARBAGE:
+def cleaning(garbage: tuple, skip: str) -> None:
+    for i in garbage:
         if i != skip:
             try:
                 FreeCAD.ActiveDocument.removeObject(i)
@@ -66,15 +72,62 @@ def cleaning(skip: str) -> None:
 
 
 def unfold(w, details: dict, path: str, skip: list = []) -> None:
-    # checking the functionality:
-    if not P.afc_additions['sm'][0]:
-        w.error.setText('Warning: SheetMetal Workbench is not available!')
-        return
 
     if len(details) == 0 or len(details) == len(skip):
         w.progress.setValue(100)
         w.status.setText('No sheet metal parts')
         return
+
+    # checking the functionality:
+    if not P.afc_additions['sm'][0]:
+        w.error.setText('Warning: SheetMetal Workbench is not available!')
+        return
+    try:
+        from SheetMetalUnfoldCmd import SMUnfoldUnattendedCommandClass as u
+    except ImportError as error:
+        P.afc_additions['sm'] = [False, '', 'color: #aa0000']
+        Logger.error(error)
+        return
+
+    # unfolder, version check:
+    if P.FC_VERSION[0] == '0' and int(P.FC_VERSION[1]) < 21:
+        new_unfolder = False
+    else:
+        if FreeCAD.ParamGet(
+                'User parameter:BaseApp/Preferences/Mod/SheetMetal').GetBool(
+                'UseOldUnfolder'):
+            new_unfolder = False
+        else:
+            new_unfolder = True
+
+    # required values:
+    if new_unfolder:
+        sketch_verification = UNFOLD_SKETCH
+        garbage = (
+            # case sensitive!
+            REPRODUCTION,
+            sketch_verification,
+            'Unfold_Sketch_Bends',
+            'Unfold_Sketch_Holes',
+            'Unfold_Sketch_Internal',
+            UNFOLD_OBJECT,  # the object must be the last
+        )
+    else:
+        sketch_verification = 'Unfold_Sketch_Outline'
+        garbage = (
+            # case sensitive!
+            REPRODUCTION,
+            UNFOLD_SKETCH,
+            'Unfold_Sketch_Bends', 'Unfold_Sketch_bends',
+            'Unfold_Sketch_Internal',
+            sketch_verification,
+            UNFOLD_OBJECT,  # the object must be the last
+        )
+
+    centering = w.checkBoxCentering.isChecked()
+    along_y = w.checkBoxAlongY.isChecked()
+
+    turn = FreeCAD.Rotation(FreeCAD.Vector(0.0, 0.0, 1.0), 90.0)
 
     steel = P.pref_steel
 
@@ -173,8 +226,7 @@ def unfold(w, details: dict, path: str, skip: list = []) -> None:
 
         # unfold:
         Logger.unfold(f'{d}: {material} ({thickness}) {k_factor}')
-        if u is not None:
-            u.Activated(None)
+        u.Activated(None)
         FreeCAD.Gui.Selection.clearSelection()
 
         # unfold, parameters:
@@ -187,9 +239,10 @@ def unfold(w, details: dict, path: str, skip: list = []) -> None:
         unfold_obj.recompute(True)
 
         # correctness check:
-        if ad.getObject(VERIFICATION_SKETCH) is None:
+        if ad.getObject(sketch_verification) is None:
+            # todo: how to check 'new_unfolder' correctly?
             Logger.warning("wrong, let's try a spare face...")
-            cleaning(REPRODUCTION)
+            cleaning(garbage, REPRODUCTION)
             # switching to spare:
             face = 'Face' + str(target[2])
             FreeCAD.Gui.Selection.addSelection(
@@ -198,44 +251,52 @@ def unfold(w, details: dict, path: str, skip: list = []) -> None:
                 u.Activated(None)
             FreeCAD.Gui.Selection.clearSelection()
             # verify:
-            if ad.getObject(VERIFICATION_SKETCH) is None:
-                cleaning('')
+            if ad.getObject(sketch_verification) is None:
+                cleaning(garbage, '')
                 Logger.error(f"'{d}' unfold error... skip")
                 continue
 
         us = ad.getObject(UNFOLD_SKETCH)
+        if us is None:
+            cleaning(garbage, '')
+            Logger.error(f"'{d}' unfold error... skip")
+            continue
+
+        us.recompute(True)
         bb = us.Shape.BoundBox
 
-        # location along the Y axis:
-        if bb.XLength < bb.YLength:
-            unfold_width = math.ceil(bb.YLength)
-            unfold_height = math.ceil(bb.XLength)
-            # centering:
-            x = bb.Center.y
-            y = bb.Center.x
-            y = abs(y) if y < 0 else -y
-            # change position:
-            r = FreeCAD.Rotation(FreeCAD.Vector(0.00, 0.00, 1.00), 90.00)
-            p = FreeCAD.Placement(FreeCAD.Vector(x, y, 0.00), r)
-            try:
-                us.Placement = p
-            except BaseException:
-                pass
-        else:
-            unfold_width = math.ceil(bb.XLength)
-            unfold_height = math.ceil(bb.YLength)
-            # centering:
-            x = bb.Center.x
-            x = abs(x) if x < 0 else -x
-            y = bb.Center.y
-            y = abs(y) if y < 0 else -y
-            # change position:
-            r = FreeCAD.Rotation(FreeCAD.Vector(0.00, 0.00, 1.00), 0.00)
-            p = FreeCAD.Placement(FreeCAD.Vector(x, y, 0.00), r)
-            try:
-                us.Placement = p
-            except BaseException:
-                pass
+        sketches = [us]
+        if new_unfolder:
+            for sketch in ('Unfold_Sketch_Holes', 'Unfold_Sketch_Internal'):
+                obj = ad.getObject(sketch)
+                if obj is not None:
+                    obj.recompute(True)
+                    if obj.Placement.Rotation.Angle == 0:
+                        sketches[0].Geometry += obj.Geometry
+                        sketches[0].recompute(True)
+                    else:
+                        along_y = False  # turning is prohibited
+                        sketches.append(obj)
+
+        if along_y and bb.XLength < bb.YLength:
+            # position along the Y axis:
+            for sketch in sketches:
+                sketch.Placement.Rotation = turn
+                sketch.recompute(True)
+            bb = sketches[0].Shape.BoundBox
+
+        x = bb.Center.x if centering else bb.XMin
+        x = abs(x) if x < 0 else -x
+        y = bb.Center.y if centering else bb.YMin
+        y = abs(y) if y < 0 else -y
+        if round(x, 6) != 0 or round(y, 6) != 0:
+            for sketch in sketches:
+                sketch.Placement.Base.x = x
+                sketch.Placement.Base.y = y
+                sketch.Placement.Base.z = 0
+                sketch.recompute(True)
+
+        bb = sketches[0].Shape.BoundBox
 
         # directory:
         target = os.path.join(path, f'{material} ({thickness})')
@@ -274,32 +335,22 @@ def unfold(w, details: dict, path: str, skip: list = []) -> None:
                     if code != '':
                         sign = f'{signature[1]}_{code}'
 
-        # checking empty signature:
-        if sign == '':
-            signature[0] = False
-
         # save:
         for i in range(int(details[d]['Quantity'])):
             if save_dxf:
                 f = os.path.join(target, f'{file} ({i + 1}).dxf')
-                importDXF.export([us], f)
-                # signature:
-                if signature[0] and len(sign) > 0:
-                    size = int(abs(unfold_width) / len(sign))
-                    size = max(6, min(size, 120))
-                    size_verify = int(abs(unfold_height) - 10)
-                    if size > size_verify:
-                        size = size_verify
-                    if P.afc_additions['ezdxf'][0]:
-                        add_signature(f, sign, unfold_width, size)
+                importDXF.export(sketches, f)
+                # postprocessor:
+                if P.afc_additions['ezdxf'][0]:
+                    dxf_postprocessor(f, sign, bb)
             if save_svg:
                 f = os.path.join(target, f'{file} ({i + 1}).svg')
-                importSVG.export([us], f)
+                importSVG.export(sketches, f)
             if save_stp:
                 f = os.path.join(target, f'{file} ({i + 1}).step')
                 ImportGui.export([body], f)
 
-        cleaning('')
+        cleaning(garbage, '')
         Logger.log('...done')
 
         progress_value += progress_step